Manual merge

+++ include/iam.h
@@ -37,12 +37,12 @@ int iam_encode_apdu(
     int segmentation,
     uint16_t vendor_id);
 
-    int iam_decode_service_request(
-        uint8_t * apdu,
-        uint32_t * pDevice_id,
-        unsigned *pMax_apdu,
-        int *pSegmentation,
-        uint16_t * pVendor_id);
+int iam_decode_service_request(
+    uint8_t * apdu,
+    uint32_t * pDevice_id,
+    uint16_t *pMax_apdu,
+    int *pSegmentation,
+    uint16_t * pVendor_id);
 
 #ifdef TEST
 #include "ctest.h"
