+++ src/iam.c
@@ -78,7 +78,7 @@ int iam_encode_apdu(
 int iam_decode_service_request(
     uint8_t * apdu,
     uint32_t * pDevice_id,
-    unsigned *pMax_apdu,
+    uint16_t *pMax_apdu,
     int *pSegmentation,
     uint16_t * pVendor_id)
 {
